RSD1MMD - Mining of Massive Datasets/league-of-legends-game-prediction/league-of-legends-spark-streaming/requirements.txt
pyspark==3.1.2
kafka-python==2.0.2
